requirements.txt
appdirs==1.4.3
dj-database-url==0.4.2
Django==1.10.6
future==0.16.0
google-api-python-client==1.6.2
gunicorn==19.7.0
httplib2==0.10.3
line-bot-sdk==1.1.0
oauth2client==4.0.0
packaging==16.8
psycopg2==2.7.1
pyasn1==0.2.3
pyasn1-modules==0.0.8
pyparsing==2.2.0
requests==2.13.0
rsa==3.4.2
six==1.10.0
uritemplate==3.0.0
whitenoise==3.3.0
